utilities_agent/requirements.txt
playwright>=1.40
pyyaml>=6.0
requests>=2.28
